refactor: Organize utility functions into appropriate packages (#203)
- Relocated common utility functions for better modularity and clarity:
  - `MultipartUtil.kt` → `core.network/MultipartUtils.kt`
  - `Constants.kt` → `common.util/Constants.kt`
  - `DateFormat.kt` → `common.util/DateExtensions.kt`
  - `daysBetween.kt` → `common.util/DateUtils.kt`
  - Consolidated:
      - `DateTimeParser.kt`, `DayOfWeek.kt`, `formatPrice.kt`, `formatRentalPeriod.kt`
        → `common.util/Formatters.kt` and `DateExtensions.kt`
- Updated all import paths across the project to match new locations
