+++ app/src/main/java/com/example/rentit/common/util/Constants.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.common
+package com.example.rentit.common.util
 
 const val D_DAY_ALERT_THRESHOLD_DAYS = 3
 const val PRICE_LIMIT = 5_000_000
