--- app/src/main/java/com/example/rentit/common/util/DateFormat.kt
@@ -1,10 +1,0 @@
-package com.example.rentit.common.util
-
-import android.os.Build
-import androidx.annotation.RequiresApi
-import java.time.LocalDate
-import java.time.format.DateTimeFormatter
-
-@RequiresApi(Build.VERSION_CODES.O)
-fun LocalDate.toShortFormat(): String =
-    this.format(DateTimeFormatter.ofPattern("yy.MM.dd"))
