--- app/src/main/java/com/example/rentit/common/util/DateTimeParser.kt
@@ -1,24 +1,0 @@
-package com.example.rentit.common.util
-
-import android.os.Build
-import androidx.annotation.RequiresApi
-import java.time.LocalDate
-import java.time.LocalDateTime
-
-@RequiresApi(Build.VERSION_CODES.O)
-fun parseLocalDateOrNull(dateStr: String): LocalDate? {
-    return try {
-        LocalDate.parse(dateStr)
-    } catch (e: Exception) {
-        null
-    }
-}
-
-@RequiresApi(Build.VERSION_CODES.O)
-fun parseLocalDateTimeOrNull(dateStr: String): LocalDateTime? {
-    return try {
-        LocalDateTime.parse(dateStr)
-    } catch (e: Exception) {
-        null
-    }
-}
