--- app/src/main/java/com/example/rentit/common/util/formatPrice.kt
@@ -1,5 +1,0 @@
-package com.example.rentit.common.util
-
-import java.text.NumberFormat
-
-fun formatPrice(price: Int): String = NumberFormat.getNumberInstance().format(price)
