+++ app/src/main/java/com/example/rentit/core/network/MultipartUtils.kt
@@ -1,0 +1,43 @@
+package com.example.rentit.core.network
+
+import android.content.Context
+import android.net.Uri
+import okhttp3.MediaType.Companion.toMediaTypeOrNull
+import okhttp3.MultipartBody
+import okhttp3.RequestBody.Companion.asRequestBody
+import java.io.File
+import java.io.FileOutputStream
+
+fun fileToMultipart(file: File): MultipartBody.Part {
+    val requestBody = file.asRequestBody("image/jpg".toMediaTypeOrNull())
+    return MultipartBody.Part.createFormData("images", file.name, requestBody)
+}
+
+fun uriToMultipart(context: Context, uri: Uri): MultipartBody.Part? {
+    val file = uriToFile(context, uri) ?: return null
+
+    val mimeType = context.contentResolver.getType(uri) ?: "image/*"
+    val requestBody = file.asRequestBody(mimeType.toMediaTypeOrNull())
+
+    return MultipartBody.Part.createFormData(
+        "thumbnailImgs",
+        file.name,
+        requestBody
+    )
+}
+
+private fun uriToFile(context: Context, uri: Uri): File? {
+    try {
+        val inputStream = context.contentResolver.openInputStream(uri)
+        val fileName = "IMG_${System.currentTimeMillis()}.jpg"
+        val tempFile = File(context.cacheDir, fileName)
+
+        inputStream?.use { input ->
+            FileOutputStream(tempFile).use { input.copyTo(it) }
+        }
+        return tempFile
+    } catch (e: Exception) {
+        e.printStackTrace()
+        return null
+    }
+}
