+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/content/owner/OwnerRentingContent.kt
@@ -22,7 +22,7 @@ import com.example.rentit.presentation.rentaldetail.components.section.RentalInf
 import com.example.rentit.presentation.rentaldetail.components.section.RentalTrackingSection
 import com.example.rentit.presentation.rentaldetail.model.RentalSummaryUiModel
 import com.example.rentit.common.theme.PrimaryBlue500
-import com.example.rentit.common.util.formatPrice
+import com.example.rentit.common.util.priceFormatter
 import com.example.rentit.domain.rental.model.RentalDetailStatusModel
 import com.example.rentit.common.enums.RentingStatus
 import kotlin.math.abs
@@ -45,7 +45,7 @@ fun OwnerRentingContent(
         )
     )
 
-    val formattedDeposit = formatPrice(rentingData.deposit)
+    val formattedDeposit = priceFormatter(rentingData.deposit)
 
     RentalInfoSection(
         title = stringResource(rentingData.status.strRes),
